poprawione 6.4 oraz 6.5

+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -2,5 +2,5 @@ package com.kodilla.testing.shape;
 
 interface Shape {
     String getShapeName();
-    int getField();
+    double getField();
 }
